Added Relaxation example.
There was no BeanFactory in the examples already. So, I added a
"SimpleFactory" that just returns fresh instances of the "_model" CFCs
when asked for.

I put the Relaxation framework files in a "com" folder and added a
"/com" mapping in Application.cfc so that this will run without
configuration.